recipes_app/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.10
django-classy-tags==2.0.0
django-cms==3.7.4
django-filer==2.0.1
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7
filelock==3.0.12
html5lib==1.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tzlocal==2.1
Unidecode==1.1.1
virtualenv==20.0.31
webencodings==0.5.1
wrapt==1.12.1
